+++ CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [1.8.9-0.6.0] - unreleased
 ### Added
 - Minecraft log file search `/moo search`
+- Analyze minions on a private island `/moo analyzeIsland` 
 - List SkyBlock info of a player `/moo stalkskyblock <playerName>`
+- Config option to toggle between Arabic and Roman numerals
+  - used for skill levels and minion tiers currently
 
 ### Changed
 - Improved handling of command error messages
@@ -50,7 +53,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Moved best friends add/remove functionality from config GUI back to commands (`/moo <add|remove> <name>`)
 - Saving best friends' UUIDs now (instead of just the name), also checking for name changes periodically to keep best friends list up to date
-
 
 ## [1.8.9-0.2.0] - 08.03.2020
 ### Added